+++ test/com/amplitude/api/UpgradePrefsTest.java
@@ -43,8 +43,6 @@ public class UpgradePrefsTest extends BaseTest {
         String sourceName = "com.amplitude.a" + "." + context.getPackageName();
         context.getSharedPreferences(sourceName, Context.MODE_PRIVATE).edit()
                 .putLong("com.amplitude.a.previousSessionId", 100L)
-                .putLong("com.amplitude.a.lastEventId", 200L)
-                .putLong("com.amplitude.a.lastEventTime", 300L)
                 .putString("com.amplitude.a.deviceId", "deviceid")
                 .putString("com.amplitude.a.userId", "userid")
                 .putBoolean("com.amplitude.a.optOut", true)
@@ -55,9 +53,6 @@ public class UpgradePrefsTest extends BaseTest {
         String targetName = Constants.PACKAGE_NAME + "." + context.getPackageName();
         SharedPreferences target = context.getSharedPreferences(targetName, Context.MODE_PRIVATE);
         assertEquals(target.getLong(Constants.PREFKEY_PREVIOUS_SESSION_ID, -1), 100L);
-        assertEquals(target.getLong(Constants.PREFKEY_LAST_EVENT_ID, -1), 200L);
-        assertEquals(target.getLong(Constants.PREFKEY_LAST_EVENT_TIME, -1), 300L);
-
         assertEquals(target.getString(Constants.PREFKEY_DEVICE_ID, null), "deviceid");
         assertEquals(target.getString(Constants.PREFKEY_USER_ID, null), "userid");
         assertEquals(target.getBoolean(Constants.PREFKEY_OPT_OUT, false), true);
@@ -94,8 +89,6 @@ public class UpgradePrefsTest extends BaseTest {
 
         String targetName = Constants.PACKAGE_NAME + "." + context.getPackageName();
         SharedPreferences target = context.getSharedPreferences(targetName, Context.MODE_PRIVATE);
-        assertEquals(target.getLong(Constants.PREFKEY_LAST_EVENT_TIME, -1), 100L);
-        assertEquals(target.getLong(Constants.PREFKEY_LAST_EVENT_ID, -1), -1);
         assertEquals(target.getLong(Constants.PREFKEY_PREVIOUS_SESSION_ID, -1), -1);
         assertEquals(target.getString(Constants.PREFKEY_DEVICE_ID, null), "deviceid");
         assertEquals(target.getString(Constants.PREFKEY_USER_ID, null), null);